chore: moving tests and fixing ci
