+++ Cargo.toml
@@ -95,7 +95,7 @@ chrono = { version = "0.4", default-features = false, features = ["now"] }
 const_format = { version = "0.2", default-features = false }
 bon = { version = "3", default-features = false }
 ## dosr
-nonstick = { version = "0.1.1", optional = true }
+nonstick = { version = "0.1", optional = true }
 libpam-sys = { version = "0.2.0", default-features = false, optional = true }
 pcre2 = { version = "0.2", default-features = false, optional = true }
 glob = { version = "0.3", default-features = false, optional = true }
