+++ src/sr/pam/mod.rs
@@ -146,10 +146,27 @@ pub(super) fn start_session<'a>(
     user: &Cred,
     cli: &Cli,
 ) -> SrResult<PamSession<impl Transaction + 'a>> {
+    let need_password_auth = if authentication.is_skip() {
+        warn!("Skipping password authentication (PAM session-only mode).");
+        false
+    } else {
+        #[cfg(feature = "timeout")]
+        {
+            !timeout::is_valid(user, user, timeout)
+        }
+        #[cfg(not(feature = "timeout"))]
+        {
+            true
+        }
+    };
+
+    debug!("need password authentication: {need_password_auth}");
+
     let conv = SrConversationHandler::builder()
         .maybe_prompt(cli.prompt.as_ref().map(|s| Cow::Owned(s.clone())))
         .use_stdin(cli.stdin)
         .build();
+
     let mut txn = TransactionBuilder::new_with_service(PAM_SERVICE)
         .username(&user.user.name)
         .build(conv.into_conversation())
@@ -158,22 +175,7 @@ pub(super) fn start_session<'a>(
             SrError::SystemError
         })?;
 
-    let skip_auth = if authentication.is_skip() {
-        warn!("Skipping authentication, this is a security risk!");
-        true
-    } else {
-        #[cfg(feature = "timeout")]
-        {
-            timeout::is_valid(user, user, timeout)
-        }
-        #[cfg(not(feature = "timeout"))]
-        {
-            false
-        }
-    };
-
-    debug!("need to re-authenticate : {}", !skip_auth);
-    if !skip_auth {
+    if need_password_auth {
         txn.authenticate(AuthnFlags::SILENT).map_err(|e| {
             error!("Authentication failed: {e}");
             SrError::AuthenticationFailed
@@ -185,14 +187,8 @@ pub(super) fn start_session<'a>(
         SrError::AuthenticationFailed
     })?;
 
-    // TODO: Enable session opening when nonstick library support it
-    // txn.open_session(AuthnFlags::SILENT).map_err(|e| {
-    //     error!("Failed to open PAM session: {}", e);
-    //     SrError::SystemError
-    // })?;
-
-    if !authentication.is_skip() {
-        #[cfg(feature = "timeout")]
+    #[cfg(feature = "timeout")]
+    {
         timeout::update_cookie(user, user, timeout).map_err(|e| {
             error!("Failed to update timeout cookie: {e}");
             SrError::SystemError
@@ -213,6 +209,7 @@ mod tests {
     };
     use serde_json::Map;
     use std::ffi::OsStr;
+    use test_log::test;
 
     // Helper function to create a test user
     fn create_test_user() -> Cred {
@@ -230,6 +227,20 @@ mod tests {
             max_usage: Some(3),
             extra_fields: Map::default(),
         }
+    }
+
+    fn pam_ready(user: &Cred) -> bool {
+        let conv = SrConversationHandler::builder().no_interact(true).build();
+
+        let mut txn = match TransactionBuilder::new_with_service(PAM_SERVICE)
+            .username(&user.user.name)
+            .build(conv.into_conversation())
+        {
+            Ok(txn) => txn,
+            Err(_) => return false,
+        };
+
+        txn.account_management(AuthnFlags::SILENT).is_ok()
     }
 
     #[test]
@@ -297,9 +308,15 @@ mod tests {
 
     #[test]
     fn test_check_auth_skip_authentication() {
+        let _ = env_logger::builder().is_test(true).try_init();
         let authentication = SAuthentication::Skip;
         let timeout = create_test_timeout();
         let user = create_test_user();
+
+        if !pam_ready(&user) {
+            eprintln!("Skipping: PAM backend not available in this test environment");
+            return;
+        }
 
         // When authentication is skipped, it should always succeed
         let cli = Cli::builder().prompt("Password: ").build();
@@ -400,6 +417,10 @@ mod tests {
         };
 
         let user = create_test_user();
+        if !pam_ready(&user) {
+            eprintln!("Skipping: PAM backend not available in this test environment");
+            return;
+        }
         let auth = SAuthentication::Skip;
 
         let cli = Cli::builder().prompt("Password: ").build();
